Move files to PolarizationCorrections folder
