+++ Framework/Algorithms/inc/MantidAlgorithms/PolarizationCorrections/HeliumAnalyserEfficiency.h
